+++ src/main/java/at/hannibal2/skyhanni/SkyHanniMod.kt
@@ -68,6 +68,8 @@ import at.hannibal2.skyhanni.features.rift.area.RiftLarva
 import at.hannibal2.skyhanni.features.rift.area.colosseum.BlobbercystsHighlight
 import at.hannibal2.skyhanni.features.rift.area.dreadfarm.RiftAgaricusCap
 import at.hannibal2.skyhanni.features.rift.area.dreadfarm.VoltHighlighter
+import at.hannibal2.skyhanni.features.rift.area.livingcave.LivingCaveDefenseBlocks
+import at.hannibal2.skyhanni.features.rift.area.livingcave.LivingCaveLivingMetalHelper
 import at.hannibal2.skyhanni.features.rift.area.livingcave.LivingMetalSuitProgress
 import at.hannibal2.skyhanni.features.rift.area.mirrorverse.DanceRoomHelper
 import at.hannibal2.skyhanni.features.rift.area.mirrorverse.RiftLavaMazeParkour
@@ -76,10 +78,7 @@ import at.hannibal2.skyhanni.features.rift.area.mirrorverse.TubulatorParkour
 import at.hannibal2.skyhanni.features.rift.area.westvillage.KloonHacking
 import at.hannibal2.skyhanni.features.rift.area.wyldwoods.RiftOdonata
 import at.hannibal2.skyhanni.features.rift.area.wyldwoods.ShyCruxWarnings
-import at.hannibal2.skyhanni.features.rift.everywhere.CruxTalismanDisplay
-import at.hannibal2.skyhanni.features.rift.everywhere.EnigmaSoulWaypoints
-import at.hannibal2.skyhanni.features.rift.everywhere.HighlightRiftGuide
-import at.hannibal2.skyhanni.features.rift.everywhere.RiftAPI
+import at.hannibal2.skyhanni.features.rift.everywhere.*
 import at.hannibal2.skyhanni.features.slayer.*
 import at.hannibal2.skyhanni.features.slayer.blaze.BlazeSlayerClearView
 import at.hannibal2.skyhanni.features.slayer.blaze.BlazeSlayerDaggerHelper
@@ -115,7 +114,7 @@ import org.apache.logging.log4j.Logger
     clientSideOnly = true,
     useMetadata = true,
     guiFactory = "at.hannibal2.skyhanni.config.ConfigGuiForgeInterop",
-    version = "0.19.Beta.8",
+    version = "0.19.Beta.10",
 )
 class SkyHanniMod {
     @Mod.EventHandler
@@ -155,6 +154,8 @@ class SkyHanniMod {
         loadModule(GardenCropMilestoneAverage())
         loadModule(GardenCropSpeed)
         loadModule(ProfileStorageData)
+        loadModule(TitleData())
+        loadModule(BlockData())
 
         // APIs
         loadModule(BazaarApi())
@@ -344,6 +345,9 @@ class SkyHanniMod {
         loadModule(LivingMetalSuitProgress())
         loadModule(VampireSlayerFeatures())
         loadModule(BlobbercystsHighlight())
+        loadModule(LivingCaveDefenseBlocks())
+        loadModule(LivingCaveLivingMetalHelper())
+        loadModule(RiftMotesOrb())
 
         init()